requirements.txt
openai
langchain
langchain-openai
python-dotenv
llama-index
chromadb